Simple setup for the system added
