+++ SereneScreen/Run.bat
@@ -1,0 +1,7 @@
+ECHO Activating conda environment...
+call conda activate serenescreen
+PAUSE
+
+ECHO Starting the App...
+streamlit run app.py  
+PAUSE
