+++ SereneScreen/runpy/setup3.py
@@ -1,0 +1,10 @@
+import subprocess
+
+def set_up_dependencies():
+    subprocess.run(["dependencies2.bat"], shell=True)
+   
+
+if __name__ == "__main__":
+    set_up_dependencies()
+ 
+
