--- app/SereneScreen/files/Untitled spreadsheet - captions (1).csv
@@ -1,2 +1,0 @@
-x,y,Summary,z,a,b,c,d
-the linux application in the linux app,"BS) o b> 6] BV & oO @ ES) File Edit Selection View Go Run Terminal EXPLORER. os v BIP Y Mental-health-analysis Y app \ screenshot > _pycache_ ® _init_py ® screenshot_app.py SereneScreen > _pycache_ > files ® apppy 1 ® utilpy 9+,M > dli Y models ® anxiety.py ® loneliness.py ® stress.py @ on tab change.ipynb @ readmemd @ screenshot_per_5_second.ipynb Y N1\Mental-health-analysis > .ipynb_checkpoints Y app > screenshot Y SereneScreen > _pycache_ v images & screenshot.png ® app.py Ei captions.csv ® capturepy > OUTLINE > TIMELINE Help ® utilpy 9+,M ® apppy 1 ® capture.py 3,M & captions.csv ..\SereneScreen M X & captions.csv ...\screenshot-test ® screenshot_app.py enh O-- N1 > Mental-health-analysis > app > SereneScreen > & captions.csv 1; PROBLEMS (i) OUTPUT DEBUG CONSOLE lstreamit +v O @ -- x TERMINAL PORTS. r window detected. window detected. window detected. window detected. window detected. window detected. window detected. window detected. window detected. window detected. active active active active active active active active active active waiting for waiting for waiting for waiting for waiting for waiting for waiting for waiting for waiting for waiting for change i change i change i change i change i change i change i change i change i change i MS SESSSEE5S 36°C ‘Smoke ENG us 1:25 PM Em Q Search eeumeasvoats & FB sro205 8 OR ","The user appears to be engaged in editing and analyzing mental health-related Python scripts, manipulating text data from a CSV file, and utilizing various tools like capturepy for capturing screenshots. They are likely managing terminal applications, observing window detections, monitoring console output via lstreamit, and possibly interacting with social media platforms through Facebook and search engines.",13:25:50,37,57,67,20
