--- app/SereneScreen/files/capt - captions.csv
@@ -1,4 +1,0 @@
-Caption,Extracted text,Summary,Time of the Event,Anxiety%,Loneliness%,Stress%,Event Time Period
-the dark mode in the dark mode menu,(© Whatsapp = Chats = 1 QP bearch or start a new chat & - Sf Ei 6/4/2024 © Euiby JM Nou: © N4-2.pdf « 18 pages 2 fe) camel squad 3:35 PM Vishwa: qrs? oO ~s Morabu - Only IITR students 2:27 PM © ~RISHABH ASATI: Obviously no ~s Sorted Rajiv A-5th Floor 2:23 PM © Oohjval: Jaha dala ho waha se le liyo Rishi_k 2:20 PM J | manually added the heading and it... Lokendra Singh Parmar 1:52 PM Okay <g> Sarthak 12:45 PM Mtlb maine to late bheja tha n5 ... Aur... x e Deepak Tailor 12:09 PM = ~ Hmm 8B Kanithi Gowtham 12:06 PM A Welcome a SENSEX eal +0:12% Q Search WhatsApp for Windows Send and receive messages without keeping your phone online. Use WhatsApp on up to 4 linked devices and 1 phone at the same time. 8 End-to-end encrypted eieumeasvroatse? a ENG us eyo 3:35 PM 6/20/2024 & @&,"The user is engaging in various activities, starting by manually adding a heading and appears to be discussing topics with different individuals. They mention an SENSEX increase and talk about WhatsApp's features on Windows devices. Later, they reference specific times for other users without explicitly detailing their actions or the context of these interactions.",15:35:29,31,50,21,15
-the visual editor in visual editor,"x File Edit Selection View Go Run Terminal Help € >| PsP J Oo og - a x | EXPLORER. ‘++ app\SereneScreen 9+,M. ® app.py Mental-health-analysis\app\SereneScreen 1 @ @ capture.py 3,M & captions.csv ...\SereneScreen M @ apppyN\.. 1MX Dy & O - v BIP N1 > Mental-health-analysis > app > SereneScreen > @ app.py >... > Mental-health-analysis 68 # Display prediction in a nice box Y N1\Mental-health-analysis 69 st.success(“Mental Health Predictior %® > ipynb_checkpoints 70 st.info(prediction) Y app m . 5 _ 72 # Add an exit button to start afresh screensho' . . a 73 if st.button(""Exit""): Y SraTESEES 74 st.experimental_rerun() - oO > _pycache_ - 75 except Exception as e: _ f © files e 76 st.error(f""Error loading csv file: {e}"") 2 Bi captions.csv uo 7 iy Ei Untitled spreadsheet - captions (1).csv U 78 # Button to start/stop the capture process i HH Untitled spreadsheet - captionscsv -U ® oF Stier (izes Cepeuns . 80 toggle_capture() A Y tress . a1| st.session_state.old_session=1 ® app.py 1M 82 B captions.csv M 83 # Display capture status ® capture.py 3,M 84 if st.session_state.capture_running: ® utilpy 85 st.info(""Capture process is running..."") > i 86 else: 87 st.info(""Capture process is stopped.”) > models eo & on tab change.ipynb BD @® readme.md 90 @ screenshot_per_5_second.ipynb a1 PROBLEMS (48) OUTPUT © DEBUGCONSOLE TERMINAL PORTS lstreamit +v () fH -- ~ x ] Anxiety%: 31 a Loneliness%: 50 stress%: 21 Screenshot Deleted: images\screenshot.png No change in active window detected. Waiting for a @ No change in active window detected. Waiting for a No change in active window detected. Waiting for a No change in active window detected. Waiting for a & > ohn No change in active window detected. Waiting for a > TIMELINE (Britt BoA Wot 81, 60135 Spaces: UTS CRLF {3 Python 3.115 (base: conda) Ae BE search aieomeasvots? ~ Mf som gan @ & ","The user is actively engaged in managing and analyzing data related to mental health, likely using an application named 'SereneScreen'. They are editing the app's configuration files, interacting with terminal commands for process management (like starting/stopping captures), monitoring session states, handling errors during CSV file processing, and observing capture statuses. Additionally, they seem focused on improving user experience by implementing features such as a prediction display box and an exit button to reset the interface.",15:35:46,30,65,74,25
-the settings menu in the dark mode,ChatGPT X | © Mental-health-analysis/cli/pipe x | @ += inpython-Google Search % | © app-Streamlit xX w app. Streamiit x + - ao xX € > G © localhost:8501 * 7 @9 | @: zi Deploy Page not found x The page that you have requested does not seem to exist. Running the app's main page. Upload a file Drag and drop file here Browse files Toggle Capture Capture process is stopped. 2 SENSEX Saal 40.12% Q Search aeoemeasvate a i e0@ goon ® O,"The user appears to be interacting with an application interface, adjusting settings preferences and possibly navigating through different functionalities such as searching or uploading files. They may have encountered technical difficulties, indicated by the message about page not found, but are attempting to access main app functions like capturing processes.",15:36:08,36,62,37,25
