--- cli/caption-testing/main.py
@@ -1,21 +1,0 @@
-from PIL import Image
-from transformers import BlipProcessor, BlipForConditionalGeneration
-
-processor = BlipProcessor.from_pretrained("Salesforce/blip-image-captioning-base")
-model = BlipForConditionalGeneration.from_pretrained("Salesforce/blip-image-captioning-base").to("cuda")
-
-img_url = r"C:\Users\Vikas\OneDrive\Desktop\BTP\Mental-health-analysis\cli\image2caption\images\screenshot_20240216_072059.png"
-raw_image = Image.open(img_url)
-
-# conditional image captioning
-text = "What is he typing or reading in the image?"
-inputs = processor(raw_image, text, return_tensors="pt").to("cuda")
-
-out = model.generate(**inputs)
-print(processor.decode(out[0], skip_special_tokens=True))
-
-# unconditional image captioning
-inputs = processor(raw_image, return_tensors="pt").to("cuda")
-
-out = model.generate(**inputs)
-print(processor.decode(out[0], skip_special_tokens=True))
